First commit for standalone combo remapper.
This commit provides a remapping of certain key combinations. The
remapping is active /everywhere/ (in all X windows and even without
X).

The next feature to be added is the ability to specify in which X
windows the remapping should be active in and which should not. (The
remapping should not be active without X.)