Add/improve comments.
This program does not map single keys to single keys/combos. It only
maps combos to single keys/combos.

A natural way to extend the program so that also [single key-->single
key/combo] maps are handled should be allowing 0 values for
mod_from. A 0 value for mod_from would mean a map for a single key.